+++ src/auto_smart_factory/include/agent/path_planning/TimedLineOfSightResult.h
@@ -6,11 +6,11 @@ class TimedLineOfSightResult {
 public:
 	bool blockedByStatic;
 	bool blockedByTimed;
-	float freeAfter;
+	double freeAfter;
 	
 	bool hasUpcomingObstacle;
-	float lastValidEntryTime;
-	float freeAfterUpcomingObstacle;
+	double lastValidEntryTime;
+	double freeAfterUpcomingObstacle;
 	
 	TimedLineOfSightResult();
 };
